DryHumidGasPvt: move implementation to impl file
